requirements.txt
charset-normalizer==3.1.0
contourpy==1.0.7
cycler==0.11.0
Deprecated==1.2.13
filelock==3.10.2
fonttools==4.39.2
idna==3.4
importlib-metadata==6.1.0
importlib-resources==5.12.0
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
llvmlite==0.39.1
MarkupSafe==2.1.2
matplotlib==3.7.1
mpmath==1.3.0
networkx==3.0
numba==0.56.4
numpy==1.23.5
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
pynndescent==0.5.8
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7.1
requests==2.28.2
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
six==1.16.0
sktime==0.16.1
sympy==1.11.1
threadpoolctl==3.1.0
torch==2.0.0
torchvision==0.15.1
tqdm==4.65.0
typing_extensions==4.5.0
umap-learn==0.5.3
urllib3==1.26.15
wrapt==1.15.0
zipp==3.15.0
